+++ src/main/java/raven/datetime/component/date/ButtonMonthYear.java
@@ -79,12 +79,19 @@ public class ButtonMonthYear extends JButton {
     }
 
     protected Color getColor() {
-        Color color = isSelected() ? UIManager.getColor("Component.accentColor") : FlatUIUtils.getParentBackground(this);
+        Color color = isSelected() ? getAccentColor() : FlatUIUtils.getParentBackground(this);
         if (press) {
             return FlatLaf.isLafDark() ? ColorFunctions.lighten(color, 0.1f) : ColorFunctions.darken(color, 0.1f);
         } else if (hover) {
             return FlatLaf.isLafDark() ? ColorFunctions.lighten(color, 0.03f) : ColorFunctions.darken(color, 0.03f);
         }
         return color;
+    }
+
+    protected Color getAccentColor() {
+        if (dateSelection.datePicker.getColor() != null) {
+            return dateSelection.datePicker.getColor();
+        }
+        return UIManager.getColor("Component.accentColor");
     }
 }